+ more nice config
